Addition and changes to the precommit hook
Changed version numbers to use the python module LooseVersion. This
module is useful for comparing version numbers in the form x.y.z and
also realizing that x.y.9 is less than x.y.10

Setup.py now performs a check to see if a pre-commit hook exists first,
and only overwrites if it does not.

version_files compares against last commit version files. This may need
to be discussed further.

+++ setup.py
@@ -7,16 +7,18 @@ import platform
 import sys
 import warnings
 import os
-
+import subprocess
 
 current_platform = platform.system().lower()
 extra_includes = []
 if current_platform == "darwin":
-    os.system("pwd")
+  if(subprocess.check_output("[ ! -f .git/hooks/pre-commit ] && echo \"Not Found\"") == "Not Found"):
     os.system("chmod +x hooks/pre-commit")
     os.system("cp -p hooks/pre-commit .git/hooks/pre-commit")
     # os.system("chmod +x hooks/pre-commit")
     # os.system("ln -s hooks/pre-commit .git/hooks/pre-commit")
+  else:
+
     try:
         import py2app
     except ImportError:
